refactor : API custiom hook 구조 팩토링
